refactor(dao, data, gin, mode): ♻️ 使用gen生成数据库模型
avt_*, cache_*, api_*

+++ utils/jwt.go
@@ -18,8 +18,8 @@ type UserInfo struct {
 }
 
 type Permissions struct {
-	Level   int `json:"level"`
-	Expires int `json:"expires"`
+	Level   int64 `json:"level"`
+	Expires int64 `json:"expires"`
 }
 
 type CustomClaims struct {
